Split into separate files and allow WithCookie and WithHeader on IOctoResponseProvider
